Today's work:
Moved exceptions.py to errors.py -- seems exceptions is already a module

Added versionMap which contains the dict that maps version numbers,
like '7.3', to the proper class.

Added the webKickstart class which a mod_python handler will be able
to call webKickstart.getKS(), pass it an IP and get a tuple containing
an error code and a string.  On error the stirng will contain the traceback
so that hopefully you can find some way of getting at it.

Did some work on other scripts.  baseKickstart is now pretty much complete
except for the bugs that I know will crop up.


git-svn-id: https://svn.linux.ncsu.edu/svn/cls/trunk/solaris2ks@956 35bbe6b0-b001-0410-9f0e-dd3f93706181