Apply similar patch to trunk:
Keys in included files aren't applied if the same key was already used in
an upper level config file.  Previous to this change one match on this
condition prevented all remaining keys in a use'd file not to be applied.

Removed some class variables in the baseKickstart class.  These variables
need to be instance variables and were probably causing some obscure
behavior.



git-svn-id: https://svn.linux.ncsu.edu/svn/cls/trunk/solaris2ks@1939 35bbe6b0-b001-0410-9f0e-dd3f93706181