TemplateVar is a class to represent each row in a metaconfig file where the first token is the keyword and the rest are considered options.
MetaTable is the API that will be exposed to the Cheetah template I think.
The main idea for both of these classes is to make creating templates
easy without the use of too much syntax for error checking, array testing,
dictionary lookups, etc.

How will errors be handled?  Right now the best way I can think of is to
do something like:

    \#silent raise WebKickstartError, "I hate text processesing"